+++ lib/bolt/bolt_option_parser.rb
@@ -31,6 +31,9 @@ module Bolt
       when 'inventory'
         { flags: OPTIONS[:inventory] + OPTIONS[:global] + %w[format inventoryfile boltdir configfile],
           banner: INVENTORY_HELP }
+      when 'group'
+        { flags: OPTIONS[:global] + %w[format inventoryfile boltdir configfile],
+          banner: GROUP_HELP }
       when 'plan'
         case action
         when 'convert'
@@ -117,6 +120,7 @@ module Bolt
         bolt secret encrypt <plaintext>  Encrypt a value
         bolt secret decrypt <encrypted>  Decrypt a value
         bolt inventory show              Show the list of targets an action would run on
+        bolt group show                  Show the list of groups in the inventory
 
       Run `bolt <subcommand> --help` to view specific examples.
 
@@ -296,6 +300,15 @@ module Bolt
 
       Available options are:
     INVENTORY_HELP
+
+    GROUP_HELP = <<~GROUP_HELP
+      Usage: bolt group <action>
+
+      Available actions are:
+        show                     Show the list of groups in the inventory
+
+      Available options are:
+    GROUP_HELP
 
     def initialize(options)
       super()